🚐 Move tests to relevant folders
